Correct review coments
