+++ QuizApp/Scenes/Home/HomeViewController.swift
@@ -11,6 +11,15 @@ import UIKit
 class HomeViewController: UIViewController {
     
     //MARK: - Components
+    private lazy var gpaView: GPAView = {
+        let gpaView = GPAView()
+        gpaView.isUserInteractionEnabled = true
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(scoreHeaderViewTapped))
+        gpaView.addGestureRecognizer(tapGesture)
+        gpaView.translatesAutoresizingMaskIntoConstraints = false
+        return gpaView
+    }()
+    
     private lazy var subjectTableView: UITableView = {
         let subjectTableView = UITableView()
         subjectTableView.separatorStyle = .none
@@ -18,61 +27,6 @@ class HomeViewController: UIViewController {
         subjectTableView.delegate = self
         subjectTableView.register(SubjectCell.self, forCellReuseIdentifier: SubjectCell.reuseIdentifier)
         return subjectTableView
-    }()
-    
-    private lazy var scoreHeaderView: UIView = {
-        let scoreHeaderView = UIView()
-        scoreHeaderView.backgroundColor = Constants.ScoreHeaderView.backgroundColor
-        scoreHeaderView.layer.cornerRadius = Constants.ScoreHeaderView.cornerRadius
-        scoreHeaderView.isUserInteractionEnabled = true
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(scoreHeaderViewTapped))
-        scoreHeaderView.addGestureRecognizer(tapGesture)
-        return scoreHeaderView
-    }()
-    
-    private let gpaView: UIView = {
-        let gpaView = UIView()
-        gpaView.backgroundColor = Constants.GpaView.backgroundColor
-        gpaView.layer.cornerRadius = Constants.GpaView.cornerRadius
-        gpaView.translatesAutoresizingMaskIntoConstraints = false
-        return gpaView
-    }()
-    
-    private let gpaTextLabel: UILabel = {
-        let gpaTextLabel = UILabel()
-        gpaTextLabel.text = Constants.GpaTextLabel.text
-        gpaTextLabel.font = Constants.GpaTextLabel.font
-        gpaTextLabel.backgroundColor = Constants.GpaTextLabel.backgroundColor
-        gpaTextLabel.textColor = Constants.GpaTextLabel.textColor
-        gpaTextLabel.translatesAutoresizingMaskIntoConstraints = false
-        return gpaTextLabel
-    }()
-    
-    private let gpaScoreLabel: UILabel = {
-        var gpaScoreLabel = UILabel()
-        gpaScoreLabel.text = Constants.GpaScoreLabel.text
-        gpaScoreLabel.font = Constants.GpaScoreLabel.font
-        gpaScoreLabel.backgroundColor = Constants.GpaScoreLabel.backgroundColor
-        gpaScoreLabel.textColor = Constants.GpaScoreLabel.textColor
-        gpaScoreLabel.translatesAutoresizingMaskIntoConstraints = false
-        return gpaScoreLabel
-    }()
-    
-    private let scoreViewDetailLabel: UILabel = {
-        let scoreViewDetailLabel = UILabel()
-        scoreViewDetailLabel.text = Constants.ScoreViewDetailLabel.text
-        scoreViewDetailLabel.font = Constants.ScoreViewDetailLabel.font
-        scoreViewDetailLabel.textColor = Constants.ScoreViewDetailLabel.textColor
-        scoreViewDetailLabel.backgroundColor = Constants.ScoreViewDetailLabel.backgroundColor
-        scoreViewDetailLabel.translatesAutoresizingMaskIntoConstraints = false
-        return scoreViewDetailLabel
-    }()
-    
-    private let scoreViewDetailImageView: UIImageView = {
-        let scoreViewDetailImageView = UIImageView()
-        scoreViewDetailImageView.image = Constants.ScoreViewDetailButton.image
-        scoreViewDetailImageView.translatesAutoresizingMaskIntoConstraints = false
-        return scoreViewDetailImageView
     }()
     
     private let chooseSubjectLabel: UILabel = {
@@ -83,6 +37,13 @@ class HomeViewController: UIViewController {
         chooseSubjectLabel.backgroundColor = Constants.ChooseSubjectLabel.backgroundColor
         chooseSubjectLabel.translatesAutoresizingMaskIntoConstraints = false
         return chooseSubjectLabel
+    }()
+    
+    private let dividerView: UIView = {
+        let dividerView = UIView()
+        dividerView.backgroundColor = Constants.DividerView.backgroundColor
+        dividerView.translatesAutoresizingMaskIntoConstraints = false
+        return dividerView
     }()
     
     private lazy var logOutButton: UIButton  = {
@@ -104,15 +65,11 @@ class HomeViewController: UIViewController {
         super.viewDidLoad()
         view.backgroundColor = .white
         setupNavigationController()
-        setupScoreHeaderViewConstraints()
-        setupGpaViewConstrains()
-        setupGpaTextLabelConstraints()
-        setupGpaScoreLabelConstraints()
-        setupScoreViewDetailLabelConstraints()
-        setupScoreViewDetailImageViewConstraints()
+        setupGPAViewConstraints()
         setupChooseSubjectLabelConstraints()
         setupTableViewConstraints()
         setupLogOutButtonConstraints()
+        setupDividerViewConstraints()
     }
     
     func setupNavigationController() {
@@ -148,6 +105,16 @@ class HomeViewController: UIViewController {
     }
     
     // MARK: - Constraints Setup
+    private func setupGPAViewConstraints() {
+        view.addSubview(gpaView)
+        NSLayoutConstraint.activate([
+            gpaView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: Constants.GPAView.topAnchor),
+            gpaView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: Constants.GPAView.leadingAnchor),
+            gpaView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: Constants.GPAView.trailingAnchor),
+            gpaView.heightAnchor.constraint(equalToConstant: Constants.GPAView.heightAnchor)
+        ])
+    }
+    
     private func setupTableViewConstraints() {
         view.addSubview(subjectTableView)
         subjectTableView.translatesAutoresizingMaskIntoConstraints = false
@@ -170,70 +137,21 @@ class HomeViewController: UIViewController {
         ])
     }
     
-    private func setupScoreHeaderViewConstraints() {
-        scoreHeaderView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(scoreHeaderView)
+    private func setupDividerViewConstraints() {
+        view.addSubview(dividerView)
         NSLayoutConstraint.activate([
-            scoreHeaderView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            scoreHeaderView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: Constants.ScoreHeaderView.leadingAnchor),
-            scoreHeaderView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: Constants.ScoreHeaderView.trailingAnchor),
-            scoreHeaderView.widthAnchor.constraint(equalToConstant: Constants.ScoreHeaderView.widthAnchor),
-            scoreHeaderView.heightAnchor.constraint(equalToConstant: Constants.ScoreHeaderView.heightAnchor)
-        ])
-    }
-    
-    private func setupGpaViewConstrains() {
-        scoreHeaderView.addSubview(gpaView)
-        NSLayoutConstraint.activate([
-            gpaView.topAnchor.constraint(equalTo: scoreHeaderView.topAnchor, constant: Constants.GpaView.topAnchor),
-            gpaView.bottomAnchor.constraint(equalTo: scoreHeaderView.bottomAnchor, constant: Constants.GpaView.bottomAnchor),
-            gpaView.leadingAnchor.constraint(equalTo: scoreHeaderView.leadingAnchor, constant: Constants.GpaView.leadingAnchor),
-            gpaView.widthAnchor.constraint(equalToConstant: Constants.GpaView.widthAnchor),
-            gpaView.heightAnchor.constraint(equalToConstant: Constants.GpaView.heightAnchor)
-        ])
-    }
-    
-    private func setupGpaTextLabelConstraints() {
-        gpaView.addSubview(gpaTextLabel)
-        NSLayoutConstraint.activate([
-            gpaTextLabel.topAnchor.constraint(equalTo: gpaView.topAnchor, constant: Constants.GpaTextLabel.topAnchor),
-            gpaTextLabel.bottomAnchor.constraint(equalTo: gpaView.bottomAnchor, constant: Constants.GpaTextLabel.bottomAnchor),
-            gpaTextLabel.leadingAnchor.constraint(equalTo: gpaView.leadingAnchor, constant: Constants.GpaTextLabel.leadingAnchor),
-        ])
-    }
-    
-    private func setupGpaScoreLabelConstraints(){
-        gpaView.addSubview(gpaScoreLabel)
-        NSLayoutConstraint.activate([
-            gpaScoreLabel.topAnchor.constraint(equalTo: gpaView.topAnchor, constant: Constants.GpaScoreLabel.topAnchor),
-            gpaScoreLabel.bottomAnchor.constraint(equalTo: gpaView.bottomAnchor, constant: Constants.GpaScoreLabel.bottomAnchor),
-            gpaScoreLabel.trailingAnchor.constraint(equalTo: gpaView.trailingAnchor, constant: Constants.GpaScoreLabel.trailingAnchor),
-            gpaScoreLabel.leadingAnchor.constraint(equalTo: gpaTextLabel.trailingAnchor,constant: Constants.GpaScoreLabel.leadingAnchor)
-        ])
-    }
-    
-    private func setupScoreViewDetailLabelConstraints() {
-        scoreHeaderView.addSubview(scoreViewDetailLabel)
-        NSLayoutConstraint.activate([
-            scoreViewDetailLabel.topAnchor.constraint(equalTo: scoreHeaderView.topAnchor, constant: Constants.ScoreViewDetailLabel.topAnchor),
-            scoreViewDetailLabel.bottomAnchor.constraint(equalTo: scoreHeaderView.bottomAnchor, constant: Constants.ScoreViewDetailLabel.bottomAnchor)
-        ])
-    }
-    
-    private func  setupScoreViewDetailImageViewConstraints() {
-        scoreHeaderView.addSubview(scoreViewDetailImageView)
-        NSLayoutConstraint.activate([
-            scoreViewDetailImageView.topAnchor.constraint(equalTo: scoreHeaderView.topAnchor, constant: Constants.ScoreViewDetailButton.topAnchor),
-            scoreViewDetailImageView.bottomAnchor.constraint(equalTo: scoreHeaderView.bottomAnchor, constant: Constants.ScoreViewDetailButton.bottomAnchor),
-            scoreViewDetailImageView.trailingAnchor.constraint(equalTo: scoreHeaderView.trailingAnchor, constant: Constants.ScoreViewDetailButton.trailingAnchor),
-            scoreViewDetailImageView.leadingAnchor.constraint(equalTo: scoreViewDetailLabel.trailingAnchor, constant: Constants.ScoreViewDetailButton.leadingAnchor)
+            dividerView.topAnchor.constraint(equalTo: subjectTableView.bottomAnchor),
+            dividerView.bottomAnchor.constraint(equalTo: logOutButton.topAnchor, constant: Constants.DividerView.bottomAnchor),
+            dividerView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            dividerView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            dividerView.heightAnchor.constraint(equalToConstant: Constants.DividerView.heightAnchor)
         ])
     }
     
     private func setupChooseSubjectLabelConstraints() {
         view.addSubview(chooseSubjectLabel)
         NSLayoutConstraint.activate([
-            chooseSubjectLabel.topAnchor.constraint(equalTo: scoreHeaderView.bottomAnchor, constant: Constants.ChooseSubjectLabel.topAnchor),
+            chooseSubjectLabel.topAnchor.constraint(equalTo: gpaView.bottomAnchor, constant: Constants.ChooseSubjectLabel.topAnchor),
             chooseSubjectLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: Constants.ChooseSubjectLabel.leadingAnchor),
         ])
     }
@@ -281,7 +199,7 @@ extension HomeViewController: UITableViewDelegate {
 private extension HomeViewController {
     enum Constants {
         
-        enum ScoreHeaderView {
+        enum GPAView {
             static let backgroundColor = UIColor(hex: "537FE7")
             static let cornerRadius: CGFloat = 26
             static let topAnchor: CGFloat = 20
@@ -289,54 +207,6 @@ private extension HomeViewController {
             static let trailingAnchor: CGFloat = -16
             static let widthAnchor: CGFloat = 344
             static let heightAnchor: CGFloat = 76
-        }
-        
-        enum GpaView {
-            static let backgroundColor = UIColor(hex: "6B91EA")
-            static let cornerRadius: CGFloat = 14
-            static let topAnchor: CGFloat = 16
-            static let bottomAnchor: CGFloat = -16
-            static let leadingAnchor: CGFloat = 18
-            static let widthAnchor: CGFloat = 88
-            static let heightAnchor: CGFloat = 40
-        }
-        
-        enum GpaTextLabel {
-            static let text = "GPA -"
-            static let backgroundColor: UIColor = .clear
-            static let textColor: UIColor = .white
-            static let font: UIFont = .systemFont(ofSize: 16)
-            static let topAnchor: CGFloat = 6
-            static let bottomAnchor: CGFloat = -6
-            static let leadingAnchor: CGFloat = 10
-        }
-        
-        enum GpaScoreLabel {
-            static let textColor = UIColor(hex: "FFD24C")
-            static let font: UIFont = .systemFont(ofSize: 16)
-            static let text = "1.0"
-            static let backgroundColor: UIColor = .clear
-            static let trailingAnchor: CGFloat = -10
-            static let topAnchor: CGFloat = 6
-            static let bottomAnchor: CGFloat = -6
-            static let leadingAnchor: CGFloat = 2
-        }
-        
-        enum ScoreViewDetailLabel {
-            static let textColor: UIColor = .white
-            static let text = "დეტალურად"
-            static let backgroundColor: UIColor = .clear
-            static let font: UIFont = .systemFont(ofSize: 16)
-            static let topAnchor: CGFloat = 28
-            static let bottomAnchor: CGFloat = -29
-        }
-        
-        enum ScoreViewDetailButton {
-            static let image = UIImage(named: "Polygon")
-            static let topAnchor: CGFloat = 30
-            static let leadingAnchor: CGFloat = 4
-            static let trailingAnchor: CGFloat = -34
-            static let bottomAnchor: CGFloat = -31
         }
         
         enum ChooseSubjectLabel {
@@ -358,7 +228,7 @@ private extension HomeViewController {
         }
         
         enum TableView {
-            static let bottomAnchor: CGFloat = -52
+            static let bottomAnchor: CGFloat = -60
         }
         
         enum TableViewCell {
@@ -370,6 +240,12 @@ private extension HomeViewController {
             static let text = "გამარჯობა,ირაკლი"
             static let font: UIFont = .systemFont(ofSize: 16)
             static let textColor = UIColor(hex: "FFC44A")
+        }
+        
+        enum DividerView {
+            static let backgroundColor = UIColor(hex: "F1F1F1")
+            static let bottomAnchor: CGFloat = -4
+            static let heightAnchor: CGFloat = 1
         }
     }
 }
