+++ QuizApp/Scenes/Quiz/QuizViewController.swift
@@ -50,10 +50,10 @@ class QuizViewController: UIViewController {
     let question: Question = Question(
         text: "რომელია ყველაზე პოპულარული პროგრამირების ენა?",
         answers: [
-            Answer(text: "Python", isCorrect: false),
-            Answer(text: "Java", isCorrect: false),
+            Answer(text: "Python"),
+            Answer(text: "Java"),
             Answer(text: "C++", isCorrect: true),
-            Answer(text: "Kotlin", isCorrect: false)
+            Answer(text: "Kotlin")
         ]
     )
     
@@ -169,21 +169,20 @@ extension QuizViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         let cellHeight = Constants.TableViewCell.cellHeight
         let distanceBetweenCells = Constants.TableViewCell.distanceBetweenCells
-        let totalHeight = cellHeight + distanceBetweenCells
-        return totalHeight
+        return cellHeight + distanceBetweenCells
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let isCorrect = question.answers[indexPath.row].isCorrect
-        
         let cell = tableView.cellForRow(at: indexPath) as? AnswerCell
-        cell?.setApperance(isCorrect: isCorrect)
+        cell?.changeBackgroundColor(isCorrect: isCorrect)
+        
         if !isCorrect {
             guard let correctIndex = question.answers.firstIndex(where: { $0.isCorrect }) else {
                 return
             }
             let correctCell = tableView.cellForRow(at: IndexPath(row: correctIndex, section: 0)) as? AnswerCell
-            correctCell?.setApperance(isCorrect: true)
+            correctCell?.changeBackgroundColor(isCorrect: true)
         }
     }
 }
